requirements.txt
streamlit
faiss-cpu
langchain
sentence-transformers
google-generativeai
